# core/aggregator/api/swagger.yaml
openapi: 3.0.0
info:
  title: Swagger Gandalf
  description: "This is a sample Petstore server.  You can find \nout more about Swagger\
    \ at \n[http://swagger.io](http://swagger.io) or on \n[irc.freenode.net, #swagger](http://swagger.io/irc/).\n"
  termsOfService: http://swagger.io/terms/
  contact:
    email: romain.fairant@orness.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0-oas3
...
servers:
- url: https://virtserver.swaggerhub.com/ditrit/Gandalf/1.0.0
- url: http://virtserver.swaggerhub.com/ditrit/Gandalf/1.0.0
...
paths:
  /role:
    get:
      tags:
      - role
      summary: List roles
      operationId: listRole
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/inline_response_200'
                x-content-type: application/json
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
    post:
      tags:
      - role
      summary: Create a role
      operationId: createRole
      requestBody:
        $ref: '#/components/requestBodies/createOrUpdateRole'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        "400":
          description: Invalid request payload
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
  /role/{roleId}:
    get:
      tags:
      - role
      summary: Find role by ID
      description: Returns a single role
      operationId: getRoleById
      parameters:
      - name: roleId
        in: path
        description: ID of role to return
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        "400":
          description: Invalid ID supplied
        "404":
          description: Role not found
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
    put:
      tags:
      - role
      summary: Updates a role
      operationId: updateRole
      parameters:
      - name: roleId
        in: path
        description: ID of role that needs to be updated
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        $ref: '#/components/requestBodies/createOrUpdateRole'
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        "400":
          description: Invalid ID supplied | Invalid request payload
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
    delete:
      tags:
      - role
      summary: Deletes a role
      operationId: deleteRole
      parameters:
      - name: roleId
        in: path
        description: Role id to delete
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Succes
        "400":
          description: Invalid ID supplied
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
...
  /tenant:
    get:
      tags:
      - tenant
      summary: List tenants
      operationId: listTenant
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/inline_response_200_2'
                x-content-type: application/json
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
    post:
      tags:
      - tenant
      summary: Create a tenant
      operationId: createTenant
      requestBody:
        $ref: '#/components/requestBodies/createOrUpdateTenant'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        "400":
          description: Invalid request payload
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
  /tenant/{tenantId}:
    get:
      tags:
      - tenant
      summary: Find tenant by ID
      description: Returns a single tenant
      operationId: getTenantById
      parameters:
      - name: tenantId
        in: path
        description: ID of tenant to return
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        "400":
          description: Invalid ID supplied
        "404":
          description: Tenant not found
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
    put:
      tags:
      - tenant
      summary: Updates a tenant
      operationId: updateTenant
      parameters:
      - name: tenantId
        in: path
        description: ID of tenant that needs to be updated
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        $ref: '#/components/requestBodies/createOrUpdateTenant'
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        "400":
          description: Invalid ID supplied | Invalid request payload
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
    delete:
      tags:
      - tenant
      summary: Deletes a tenant
      operationId: deleteTenant
      parameters:
      - name: tenantId
        in: path
        description: Tenant id to delete
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Succes
        "400":
          description: Invalid ID supplied
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
  /secretAssignement:
    get:
      tags:
      - secretAssignement
      summary: List SecretAssignements
      operationId: listSecretAssignement
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/inline_response_200_3'
                x-content-type: application/json
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
    post:
      tags:
      - secretAssignement
      summary: Create a SecretAssignement
      operationId: createSecretAssignement
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecretAssignement'
        "400":
          description: Invalid request payload
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
...
components:
  schemas:
    Model:
      type: object
      properties:
        id:
          type: integer
          description: The ID.
        createdAt:
          type: string
          description: The created date.
          format: date-time
        updatedAt:
          type: string
          description: The updated date.
          format: date-time
        deletedAt:
          type: string
          description: The deleted date.
          format: date-time
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        deletedAt: 2000-01-23T04:56:07.000+00:00
        id: 0
        updatedAt: 2000-01-23T04:56:07.000+00:00
...
    Role:
      type: object
      properties:
        model:
          $ref: '#/components/schemas/Model'
        name:
          type: string
          description: The role name.
      example:
        name: name
        model:
          createdAt: 2000-01-23T04:56:07.000+00:00
          deletedAt: 2000-01-23T04:56:07.000+00:00
          id: 0
          updatedAt: 2000-01-23T04:56:07.000+00:00
...
    Tenant:
      type: object
      properties:
        model:
          $ref: '#/components/schemas/Model'
        name:
          type: string
          description: The tenant name.
        password:
          type: string
          description: The tenant password.
      example:
        password: password
        name: name
        model:
          createdAt: 2000-01-23T04:56:07.000+00:00
          deletedAt: 2000-01-23T04:56:07.000+00:00
          id: 0
          updatedAt: 2000-01-23T04:56:07.000+00:00
    SecretAssignement:
      type: object
      properties:
        secret:
          type: string
          description: The secret.
        addressIP:
          type: string
          description: The secret IP Address.
      example:
        secret: secret
        addressIP: addressIP
...
    inline_response_200:
      type: object
      properties:
        schema:
          $ref: '#/components/schemas/Role'
      example:
        schema:
          name: name
          model:
            createdAt: 2000-01-23T04:56:07.000+00:00
            deletedAt: 2000-01-23T04:56:07.000+00:00
            id: 0
            updatedAt: 2000-01-23T04:56:07.000+00:00
...
    inline_response_200_2:
      type: object
      properties:
        schema:
          $ref: '#/components/schemas/Tenant'
      example:
        schema:
          password: password
          name: name
          model:
            createdAt: 2000-01-23T04:56:07.000+00:00
            deletedAt: 2000-01-23T04:56:07.000+00:00
            id: 0
            updatedAt: 2000-01-23T04:56:07.000+00:00
    inline_response_200_3:
      type: object
      properties:
        schema:
          $ref: '#/components/schemas/SecretAssignement'
      example:
        schema:
          secret: secret
          addressIP: addressIP
...
  requestBodies:
    createOrUpdateRole:
      description: The role to create.
      content:
        application/json:
          schema:
            required:
            - name
            type: object
            properties:
              name:
                type: string
...
    createOrUpdateTenant:
      description: The tenant to create.
      content:
        application/json:
          schema:
            required:
            - name
            type: object
            properties:
              name:
                type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT